+++ projects/pip-webui2-nav/src/lib/nav-header/nav-header.component.spec.ts
@@ -1,35 +1,35 @@
-import { TestBed, async } from '@angular/core/testing';
+// import { TestBed, async } from '@angular/core/testing';
 
-import { MatToolbarModule } from '@angular/material';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { CommonModule } from '@angular/common';
-import { BrowserModule } from '@angular/platform-browser';
-import { PipNavHeaderComponent } from './nav-header.component';
+// import { MatToolbarModule } from '@angular/material';
+// import { FlexLayoutModule } from '@angular/flex-layout';
+// import { CommonModule } from '@angular/common';
+// import { BrowserModule } from '@angular/platform-browser';
+// import { PipNavHeaderComponent } from './nav-header.component';
 
-import { PipNavPartService } from '../navpart/shared/navpart.service';
+// import { PipNavPartService } from '../navpart/shared/navpart.service';
 
 describe('PipNavHeaderComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        MatToolbarModule,
-        MatToolbarModule,
-        FlexLayoutModule,
-        CommonModule,
-        BrowserModule
-      ],
-      declarations: [
-        PipNavHeaderComponent
-      ],
-      providers: [
-        PipNavPartService
-      ]
-    }).compileComponents();
-  }));
+  // beforeEach(async(() => {
+  //   TestBed.configureTestingModule({
+  //     imports: [
+  //       MatToolbarModule,
+  //       MatToolbarModule,
+  //       FlexLayoutModule,
+  //       CommonModule,
+  //       BrowserModule
+  //     ],
+  //     declarations: [
+  //       PipNavHeaderComponent
+  //     ],
+  //     providers: [
+  //       PipNavPartService
+  //     ]
+  //   }).compileComponents();
+  // }));
 
-  it('should create the nav-header', async(() => {
-    const fixture = TestBed.createComponent(PipNavHeaderComponent);
-    const component = fixture.debugElement.componentInstance;
-    expect(component).toBeTruthy();
-  }));
+  // it('should create the nav-header', async(() => {
+  //   const fixture = TestBed.createComponent(PipNavHeaderComponent);
+  //   const component = fixture.debugElement.componentInstance;
+  //   expect(component).toBeTruthy();
+  // }));
 });