+++ src/app/navparts-example/navparts-example.module.ts
@@ -15,7 +15,7 @@ import { NavpartsExampleComponent } from './navparts-example.component';
     FlexLayoutModule,
     FormsModule, ReactiveFormsModule,
     MatButtonModule, MatInputModule, MatToolbarModule, MatListModule,
-    TranslateModule,
+    TranslateModule.forChild(),
     PipDocumentLayoutModule, PipShadowModule,
     PipNavModule
   ],