+++ CHANGELOG.md
@@ -57,3 +57,23 @@ No breaking changes in this version
 
 ### Improvements
 Event for selection item in **nav menu**
+
+## <a name="1.1.0"></a> 1.1.0 (2017-11-22)
+
+Update to stable Angular 7 and fix the tests.
+
+### Features
+* **nav icon**: main app icon
+* **nav menu**: main navigation menu with list of links
+* **nav header**: navigation toolbar that contains key nav information
+* **breadcrumb**: contains title of each section where was the user
+* **primary actions**: list of primary links that places in app bar
+* **secondary actions**: menu of secondry links
+
+### Breaking Changes
+No breaking changes in this version
+
+### Bug Fixes
+* Library tests works now. Run `npm run test:lib` to check it out;
+* Example tests works now. Run `npm run test` to check it out;
+* E2E tests works too. Run `npm run e2e` to check it out.